fix(getCustomer): nuMsisdnAlternativo Int->Long + bump 2.0.1
MSISDN alternativo é telefone; como Int estourava (>2.1bi) e crashava o parse
quando preenchido. v2.0.0 não disparava só pq a conta testada tinha o campo nulo.

+++ shared/src/main/kotlin/io/github/surfdevops/surfapikit/features/customer/getcustomer/GetCustomer.kt
@@ -29,7 +29,7 @@ data class GetCustomerSuccess(
         val tpDocumento: String? = null,
         val dsPassword: String,
         val dsEmail: String? = null,
-        val nuMsisdnAlternativo: Int? = null,
+        val nuMsisdnAlternativo: Long? = null,
         val dsCEP: String? = null,
         val dsEndereco: String? = null,
         val dsNumero: Int? = null,
